Inicio de angular

+++ src/app/app.component.ts
@@ -2,12 +2,13 @@ import { Router } from '@angular/router';
 import { MatDialog } from '@angular/material';
 import {MediaMatcher} from '@angular/cdk/layout';
 import {MatSidenav} from '@angular/material/sidenav';
-import {ViewChild, ChangeDetectorRef, Component, OnDestroy, OnInit} from '@angular/core';
+import {ViewChild, ChangeDetectorRef, Component, OnDestroy, OnInit, Inject, PLATFORM_ID} from '@angular/core';
 import { LoginComponent } from './components/modals/login/login.component';
 import { SignupComponent } from './components/modals/signup/signup.component';
 import { ShareLoginService } from './services/shareLogin.service';
 import { AuthService } from './services/auth.service';
 import { User } from './models/user';
+import { isPlatformBrowser } from '@angular/common';
 
 @Component({
   selector: 'app-root',
@@ -25,6 +26,7 @@ export class AppComponent implements OnDestroy, OnInit {
 
   constructor(changeDetectorRef: ChangeDetectorRef,
               media: MediaMatcher,
+              @Inject(PLATFORM_ID) private platformId: any,
               private router: Router,
               public dialog: MatDialog,
               public authService: AuthService,
@@ -35,12 +37,15 @@ export class AppComponent implements OnDestroy, OnInit {
   }
 
   ngOnInit(): void {
-    const token: string = localStorage.getItem('token');
-    // tslint:disable-next-line: triple-equals
-    if (token != '') {
-      this.shareLoginService.sendLogin(this.authService.isAuthenticated());
-      this.shareLoginService.sendUser(this.authService.getRol());
+    if (isPlatformBrowser(this.platformId)) {
+      const token: string = localStorage.getItem('token');
+      // tslint:disable-next-line: triple-equals
+      if (token != '') {
+        this.shareLoginService.sendLogin(this.authService.isAuthenticated());
+        this.shareLoginService.sendUser(this.authService.getRol());
+      }
     }
+
     this.shareLoginService.loggedSource.subscribe(logged => {
       this.logged = this.authService.isAuthenticated();
     });
@@ -50,7 +55,9 @@ export class AppComponent implements OnDestroy, OnInit {
   }
 
   logout() {
-    localStorage.clear();
+    if (isPlatformBrowser(this.platformId)) {
+      localStorage.clear();
+    }
     this.shareLoginService.sendLogin(false);
     this.shareLoginService.sendUser('');
     this.snav.close();
